force constants to ull

+++ c/003.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <math.h>
 
-void set_max_prime(const uint64_t big, int* max_prime_ptr)
+void set_max_prime(const unsigned long long big, int* max_prime_ptr)
 {
   int is_prime = 1;
   int count = 3;
@@ -18,14 +18,14 @@ void set_max_prime(const uint64_t big, int* max_prime_ptr)
 
 int main()
 {
-  uint64_t number = 600851475143;
+  unsigned long long number = 600851475143ull;
   int root = 775147; // the first odd integer larger than root(number)
 
   int max_prime = 0;
   int i;
   for (i=root; i>2; i=i-2) // just use odd factors
   {
-    uint64_t divi = 0;
+    unsigned long long divi = 0;
     if (!(number % i))
     {
       divi = number/i;
